+++ src/emulator/utils.h
@@ -16,12 +16,12 @@ void loadBinary(char* memory, char* path);
 int getWord(char* memory);
 
 // Rotate right operations
-int rotateRight64(long long int value, int shift);
-int rotateRight32(long long int value, int shift);
+int rotateRight64(uint64_t value, int shift);
+int rotateRight32(uint64_t value, int shift);
 
 // Arithmetic shift operations
-int arithmeticShiftRight64(long long int value, int shift);
-int arithmeticShiftRight32(long long int value, int shift);
+int arithmeticShiftRight64(uint64_t value, int shift);
+int arithmeticShiftRight32(uint64_t value, int shift);
 
 // Gets bit at kth position from n.
 int getBitAt(int n, int k);
